@@ -18,72 +21,70 @@
   "main": "dist/index.js",
   "homepage": "https://bitbucket.org/Skiminojek/widget-camera/src",
   "devDependencies": {
-    "@babel/core": "^7.3.3",
-    "@babel/plugin-proposal-class-properties": "^7.3.3",
-    "@babel/plugin-transform-async-to-generator": "^7.2.0",
-    "@babel/plugin-transform-react-jsx": "^7.3.0",
-    "@babel/plugin-transform-runtime": "^7.2.0",
-    "@babel/preset-env": "^7.3.1",
-    "autoprefixer": "^9.4.8",
-    "babel-eslint": "^10.0.1",
-    "babel-loader": "^8.0.5",
+    "@babel/core": "^7.8.4",
+    "@babel/plugin-proposal-class-properties": "^7.8.3",
+    "@babel/plugin-transform-async-to-generator": "^7.8.3",
+    "@babel/plugin-transform-react-jsx": "^7.8.3",
+    "@babel/plugin-transform-runtime": "^7.8.3",
+    "@babel/preset-env": "^7.8.4",
+    "autoprefixer": "^9.7.4",
+    "babel-eslint": "^10.0.3",
+    "babel-loader": "^8.0.6",
     "breakpoint-sass": "^2.7.1",
     "chai": "^4.2.0",
-    "clean-webpack-plugin": "^1.0.1",
-    "copy-webpack-plugin": "^4.6.0",
+    "clean-webpack-plugin": "^3.0.0",
+    "copy-webpack-plugin": "^5.1.1",
     "cross-env": "^7.0.0",
-    "css-loader": "^2.1.0",
+    "css-loader": "^3.4.2",
     "cssnano": "^4.1.10",
-    "eslint": "^5.14.1",
-    "eslint-config-airbnb": "^17.1.0",
-    "eslint-plugin-import": "^2.16.0",
-    "eslint-plugin-jsx-a11y": "^6.2.1",
-    "eslint-plugin-react": "^7.12.4",
-    "file-loader": "^3.0.1",
+    "eslint": "^6.8.0",
+    "eslint-config-airbnb": "^18.0.1",
+    "eslint-plugin-import": "^2.20.1",
+    "eslint-plugin-jsx-a11y": "^6.2.3",
+    "eslint-plugin-react": "^7.18.3",
+    "file-loader": "^5.0.2",
     "git-revision-webpack-plugin": "^3.0.4",
     "html-loader": "^0.5.5",
     "html-webpack-inline-source-plugin": "0.0.10",
     "html-webpack-plugin": "^3.2.0",
-    "image-webpack-loader": "^4.6.0",
-    "jest": "^24.1.0",
+    "image-webpack-loader": "^6.0.0",
+    "jest": "^25.1.0",
     "jsdoc": "^3.6.3",
-    "jsdoc-to-markdown": "^4.0.1",
-    "karma": "^4.2.0",
-    "karma-chrome-launcher": "^2.2.0",
-    "karma-coverage": "^1.1.2",
+    "jsdoc-to-markdown": "^5.0.3",
+    "karma": "^4.4.1",
+    "karma-chrome-launcher": "^3.1.0",
+    "karma-coverage": "^2.0.1",
     "karma-mocha": "^1.3.0",
     "karma-sourcemap-loader": "^0.3.7",
     "karma-spec-reporter": "0.0.32",
-    "karma-webpack": "^3.0.5",
+    "karma-webpack": "^4.0.2",
     "minami": "^1.2.3",
     "mini-css-extract-plugin": "^0.9.0",
     "mocha": "^6.1.4",
-    "node-sass": "^4.11.0",
+    "node-sass": "^4.13.1",
     "postcss-loader": "^3.0.0",
     "preact-compat-enzyme": "^0.2.5",
-    "preact-jsx-chai": "^2.2.1",
     "preact-test-utils": "^0.1.4",
-    "puppeteer": "^1.18.1",
-    "regenerator-runtime": "^0.13.1",
+    "puppeteer": "^2.1.1",
+    "regenerator-runtime": "^0.13.3",
     "rename-output-webpack-plugin": "^1.0.1",
-    "sass-loader": "^7.1.0",
-    "sinon": "^7.2.4",
-    "standard-version": "^6.0.1",
-    "style-loader": "^0.23.1",
-    "terser-webpack-plugin": "^1.2.2",
-    "url-loader": "^3.0.0",
-    "webpack": "^4.29.5",
+    "sass-loader": "^8.0.2",
+    "sinon": "^8.1.1",
+    "standard-version": "^7.1.0",
+    "style-loader": "^1.1.3",
+    "terser-webpack-plugin": "^2.3.4",
+    "webpack": "^4.41.5",
     "webpack-clean": "^1.2.3",
     "webpack-cli": "^3.2.3",
-    "webpack-dev-server": "^3.1.14"
+    "webpack-dev-server": "^3.10.3"
   },
   "dependencies": {
-    "@babel/polyfill": "^7.2.5",
-    "@babel/runtime": "^7.3.1",
+    "@babel/polyfill": "^7.8.3",
+    "@babel/runtime": "^7.8.4",
     "classnames": "^2.2.6",
-    "preact": "^8.4.2"
+    "preact": "^10.3.1"
   },
   "peerDependencies": {
-    "preact": "^8.4.2"
+    "preact": "^10.3.1"
   }
 }
